% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP_data_prepration.R
\name{ICP_climate_soil}
\alias{ICP_climate_soil}
\title{Prepare the ICP climate and soil database
This function reads the ISRIC soil data from ICP files that should be provided and processes the data to create processed dataframes for climate, soil, stand and thinning data from ICP database.}
\usage{
ICP_climate_soil(
  icp_climate = NULL,
  chelsa_pr = NULL,
  chelsa_tas = NULL,
  chelsa_tas_min = NULL,
  chelsa_tas_max = NULL,
  chelsa_rad = NULL,
  so_pfh = NULL,
  so_som = NULL,
  ss_ssm = NULL,
  si_sta = NULL,
  gr_ipm = NULL,
  gr_iev = NULL,
  gr_inv = NULL,
  lf_lfm = NULL
)
}
\arguments{
\item{ICP_temperature:}{Temperature parameters from ICP}

\item{ICP_precipitation:}{Precipitation parameter from ICP}

\item{ICP_radation:}{Radiation parameter from ICP}

\item{country_names:}{Country names}

\item{country_codes:}{Country codes}

\item{country_lat_lon:}{Country lat and lon}

\item{chelsa_clim:}{Climate parameters from CHELSA}

\item{soil_data:}{ICP soil data}

\item{soil_som_data:}{ICP soil SOM data}

\item{soil_ssm_data:}{ICP soil SOM data}

\item{stand_data:}{ICP Stand data}

\item{ss_age_d_1:}{ICP stand age data Method 1}

\item{ss_age_d_2:}{ICP stand age data Method 2}

\item{tree_data_dia:}{ICP tree diameter Method 1}

\item{thinning_data:}{ICP thinning data}

\item{inventory_data:}{ICP tree inventory data}

\item{litter_data:}{ICP litter data}

\item{frosster_d:}{Frosster et al. parameters}
}
\value{
A list containing processed dataframes for soil and climate data from ICP database
}
\description{
Prepare the ICP climate and soil database
This function reads the ISRIC soil data from ICP files that should be provided and processes the data to create processed dataframes for climate, soil, stand and thinning data from ICP database.
}
\examples{
ICP_climate_soil(icp_climate, chelsa_pr, chelsa_tas, chelsa_tas_min, chelsa_tas_max, chelsa_rad, so_pfh, so_som,ss_ssm, si_sta, gr_ipm, gr_iev, gr_inv, lf_lfm)
}
