% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIASA_ICP_data.R
\name{IIASA_ICP_process}
\alias{IIASA_ICP_process}
\title{Process the ICP climate and soil database
This function process the loaded ICP,ISRIC and CHELSA data for model evalution/calibration or other purposes}
\usage{
IIASA_ICP_process(
  isric_soil_database = NULL,
  icp_database = NULL,
  country = NULL,
  site_number = NULL
)
}
\arguments{
\item{climate:}{Climate data output}

\item{soil:}{Site soil data output}

\item{species:}{Site species data output}

\item{thinning:}{Site thinning data output}
}
\value{
A list containing processed data frames for soil and climate data from all the loaded database
}
\description{
Process the ICP climate and soil database
This function process the loaded ICP,ISRIC and CHELSA data for model evalution/calibration or other purposes
}
\examples{
 IIASA_ICP_process(isric_soil_d, icp_data, site_number)
}
