% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_data_prepration.R
\name{isric_soil_data}
\alias{isric_soil_data}
\title{Process ISRIC soil data to create processed dataframes for soil data and available water capacity (AWC).}
\usage{
isric_soil_data(
  awc_files = NULL,
  clay_files = NULL,
  silt_files = NULL,
  sand_files = NULL,
  soc_files = NULL,
  nitrogen_files = NULL
)
}
\arguments{
\item{soil_data}{A null object (unused).}

\item{isric_data_awc}{A dataframe containing longitude, latitude, and available water capacity (AWC) data for each point (must contain columns named "longitude", "latitude", and "awc").}

\item{isric_data_soil}{A dataframe containing longitude, latitude, and soil attribute data (clay, silt, sand, and  SOC) for each point (must contain columns named "longitude", "latitude", "clay", "silt", "sand", and "soc").}

\item{isric_data_nitrogen}{A dataframe containing longitude, latitude, and soil attribute data (Nitrogen) for each point (must contain columns named "longitude", "latitude", "nitrogen").}
}
\value{
A list containing two processed dataframes: "soil_data" and "awc_data". The "soil_data" dataframe contains processed soil attribute data (clay, silt, sand, SOC, and Nitrogen) for each point. The "awc_data" dataframe contains processed available water capacity (AWC) data for each point.
}
\description{
This function reads the ISRIC soil data from raster files and processes the data to create processed dataframes for soil data and available water capacity (AWC).
}
\examples{
isric_soil_data(awc_files, clay_files, silt_files, sand_files, soc_files, nitrogen_files)

}
